ChatBot/requirements.txt
Flask
psycopg2-binary
spacy
